Simplify the usage of boundCallback (now partial) and fix caching bug
Much simpler to use the createCachedPartial function to create a method that can be used with context and the cache already there. This way `boundCallback(this, this.boundCallbacks, this.handler, arg1, arg2)` now becomes `this.partial(this.handler, arg1, arg2)` and we just have to add `this.partial = createCachedPartial(this)` in the constructor.

The caching bug was that boundCallback only used the keys to create a cache, but the cache is shared by many functions and so it should use the function too. Ideally we could use the function and the keys straight up as you can with ImmutableJS since it supports any object as keys into a Map. Unfortunately plain JS objects have to be keyed by string/number, so we compromise by taking the signature of the function only and trust the user to be wary.